requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1
django-cleanup==8.1.0
django-filter==24.3
django-iranian-cities==1.0.2
django-jalali==7.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.7.0

factory_boy~=3.3.1
Faker~=28.4.1